requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.1
mysql-connector-python==9.3.0
sqlparse==0.5.3
tzdata==2025.2
